fix: derive SamlConfig from SAMLOptions (#515)
* refactor: reorder SAMLOptions types to compare with SamlConfig

* fix: expand types from SAMLOptions

* fix: derive SamlConfig from SAMLOptions
